rename package pkg/linksharing to linksharing
